+++ src/components/Textarea/Textarea.types.ts
@@ -5,5 +5,6 @@ export interface TextareaProps {
   error?: string
   placeholder?: string
   rows?: number
+  name: string
   onChange?: ChangeEventHandler<HTMLTextAreaElement>
 }