+++ src/stories/Textarea.stories.tsx
@@ -10,10 +10,10 @@ export default {
 const Template: ComponentStory<typeof Textarea> = (args) => <Textarea {...args} />
 
 export const Default = Template.bind({})
-Default.args = { placeholder: 'testar med text' }
+Default.args = { placeholder: 'testar med text', name: 'Textarea' }
 
 export const Disable = Template.bind({})
-Disable.args = { disabled: true, placeholder: 'testar med text' }
+Disable.args = { disabled: true, placeholder: 'testar med text', name: 'TextareaDisabled' }
 
 export const Rows = Template.bind({})
-Rows.args = { rows: 10, placeholder: 'testar med text' }
+Rows.args = { rows: 10, placeholder: 'testar med text', name: 'TextareaRows' }