+++ CHANGELOG.md
@@ -6,20 +6,29 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
 ### Fixed
+
 - Add peerDependencies for @tailwindcss/forms
 - Add name to Input-component
+- Add name to Textarea-component
 
 ## 1.1.7 - 2022-12-14
+
 ### Fixed
+
 - Updated classes on Input-component
 - Updated classes on Textarea-component
 
 ## 1.1.6 - 2022-12-14
+
 ### Fixed
+
 - Export InputType
 
 ## 1.1.3 - 2022-12-14
+
 ### Added
+
 - Changelog
 - An icon props on Input-component